refactor: move ethereum and op builder into separate crates
