+++ bin/reth/Cargo.toml
@@ -43,6 +43,7 @@ reth-downloaders = { workspace = true, features = ["test-utils"] }
 reth-tracing.workspace = true
 reth-tasks.workspace = true
 reth-net-nat.workspace = true
+reth-optimism-payload-builder = { workspace = true, optional = true }
 reth-payload-builder.workspace = true
 reth-payload-validator.workspace = true
 reth-basic-payload-builder.workspace = true
@@ -133,7 +134,7 @@ optimism = [
     "reth-provider/optimism",
     "reth-beacon-consensus/optimism",
     "reth-auto-seal-consensus/optimism",
-    "reth-basic-payload-builder/optimism",
+    "reth-optimism-payload-builder",
     "reth-network/optimism",
     "reth-network-api/optimism",
     "reth-blockchain-tree/optimism",
